- [ZBX-3730] Update to latest trunk
git-svn-id: svn://svn.zabbix.com/branches/dev@19666 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-3730/frontends/php/api/classes/class.cuser.php
@@ -35,8 +35,8 @@ class CUser extends CZBXAPI{
  * @param array $options['usrgrpids'] filter by UserGroup IDs
  * @param array $options['userids'] filter by User IDs
  * @param boolean $options['type'] filter by User type [ USER_TYPE_ZABBIX_USER: 1, USER_TYPE_ZABBIX_ADMIN: 2, USER_TYPE_SUPER_ADMIN: 3 ]
- * @param boolean $options['select_usrgrps'] extend with UserGroups data for each User
- * @param boolean $options['get_access'] extend with access data for each User
+ * @param boolean $options['selectUsrgrps'] extend with UserGroups data for each User
+ * @param boolean $options['getAccess'] extend with access data for each User
  * @param boolean $options['count'] output only count of objects in result. ( result returned in property 'rowscount' )
  * @param string $options['pattern'] filter by Host name containing only give pattern
  * @param int $options['limit'] output will be limited to given number
@@ -47,7 +47,6 @@ class CUser extends CZBXAPI{
 	public function get($options=array()){
 		$result = array();
 		$user_type = self::$userData['type'];
-		$userid = self::$userData['userid'];
 
 		$sort_columns = array('userid', 'alias'); // allowed columns for sorting
 		$subselects_allowed_outputs = array(API_OUTPUT_REFER, API_OUTPUT_EXTEND); // allowed output options for [ select_* ] params
@@ -69,17 +68,17 @@ class CUser extends CZBXAPI{
 // filter
 			'filter'					=> null,
 			'search'					=> null,
-			'searchByAny'			=> null,
+			'searchByAny'				=> null,
 			'startSearch'				=> null,
 			'excludeSearch'				=> null,
 
 // OutPut
 			'output'					=> API_OUTPUT_REFER,
 			'editable'					=> null,
-			'select_usrgrps'			=> null,
-			'select_medias'				=> null,
-			'select_mediatypes'			=> null,
-			'get_access'				=> null,
+			'selectUsrgrps'				=> null,
+			'selectMedias'				=> null,
+			'selectMediatypes'			=> null,
+			'getAccess'					=> null,
 			'countOutput'				=> null,
 			'preservekeys'				=> null,
 
@@ -97,7 +96,7 @@ class CUser extends CZBXAPI{
 			$sql_parts['select']['userid'] = ' u.userid';
 			foreach($options['output'] as $key => $field){
 				if(isset($dbTable['fields'][$field]))
-					$sql_parts['select'][$field] = ' u.'.$field;
+					$sql_parts['select'][$field] = 'u.'.$field;
 			}
 
 			$options['output'] = API_OUTPUT_CUSTOM;
@@ -246,12 +245,12 @@ class CUser extends CZBXAPI{
 				else{
 					if(!isset($result[$user['userid']])) $result[$user['userid']]= array();
 
-					if($options['select_usrgrps'] && !isset($result[$user['userid']]['usrgrps'])){
+					if($options['selectUsrgrps'] && !isset($result[$user['userid']]['usrgrps'])){
 						$result[$user['userid']]['usrgrps'] = array();
 					}
 
 // usrgrpids
-					if(isset($user['usrgrpid']) && is_null($options['select_usrgrps'])){
+					if(isset($user['usrgrpid']) && is_null($options['selectUsrgrps'])){
 						if(!isset($result[$user['userid']]['usrgrps']))
 							$result[$user['userid']]['usrgrps'] = array();
 
@@ -260,7 +259,7 @@ class CUser extends CZBXAPI{
 					}
 
 // mediaids
-					if(isset($user['mediaid']) && is_null($options['select_medias'])){
+					if(isset($user['mediaid']) && is_null($options['selectMedias'])){
 						if(!isset($result[$user['userid']]['medias']))
 							$result[$user['userid']]['medias'] = array();
 
@@ -269,7 +268,7 @@ class CUser extends CZBXAPI{
 					}
 
 // mediatypeids
-					if(isset($user['mediatypeid']) && is_null($options['select_mediatypes'])){
+					if(isset($user['mediatypeid']) && is_null($options['selectMediatypes'])){
 						if(!isset($result[$user['userid']]['mediatypes']))
 							$result[$user['userid']]['mediatypes'] = array();
 
@@ -287,7 +286,7 @@ Copt::memoryPick();
 		}
 
 // Adding Objects
-		if(!is_null($options['get_access'])){
+		if(!is_null($options['getAccess'])){
 			foreach($result as $userid => $user){
 				$result[$userid] += array('gui_access' => 0, 'debug_mode' => 0, 'users_status' => 0);
 			}
@@ -305,17 +304,17 @@ Copt::memoryPick();
 		}
 
 // Adding usergroups
-		if(!is_null($options['select_usrgrps']) && str_in_array($options['select_usrgrps'], $subselects_allowed_outputs)){
+		if(!is_null($options['selectUsrgrps']) && str_in_array($options['selectUsrgrps'], $subselects_allowed_outputs)){
 			$obj_params = array(
-				'output' => $options['select_usrgrps'],
+				'output' => $options['selectUsrgrps'],
 				'userids' => $userids,
-				'preservekeys' => 1
+				'preservekeys' => true
 			);
 			$usrgrps = API::UserGroup()->get($obj_params);
-			foreach($usrgrps as $usrgrpid => $usrgrp){
+			foreach($usrgrps as $usrgrp){
 				$uusers = $usrgrp['users'];
 				unset($usrgrp['users']);
-				foreach($uusers as $num => $user){
+				foreach($uusers as $user){
 					$result[$user['userid']]['usrgrps'][] = $usrgrp;
 				}
 			}
@@ -323,10 +322,20 @@ Copt::memoryPick();
 
 // TODO:
 // Adding medias
-		if(!is_null($options['select_medias']) && str_in_array($options['select_medias'], $subselects_allowed_outputs)){
+		if(!is_null($options['selectMedias']) && str_in_array($options['selectMedias'], $subselects_allowed_outputs)){
+			$obj_params = array(
+				'output' => $options['selectMedias'],
+				'userids' => $userids,
+				'preservekeys' => true
+			);
+			$userMedias = API::UserMedia()->get($obj_params);
+			foreach($userMedias as $mediaid => $media){
+				$result[$media['userid']]['medias'][] = $media;
+			}
 		}
+
 // Adding mediatypes
-		if(!is_null($options['select_mediatypes']) && str_in_array($options['select_mediatypes'], $subselects_allowed_outputs)){
+		if(!is_null($options['selectMediatypes']) && str_in_array($options['selectMediatypes'], $subselects_allowed_outputs)){
 		}
 
 // removing keys (hash -> array)
@@ -371,6 +380,9 @@ Copt::memoryPick();
 				$dbUser = $user;
 			}
 			else if($update){
+				if(!isset($dbUsers[$user['userid']]))
+					self::exception(ZBX_API_ERROR_PARAMETERS, _s('You do not have permissions to update user or user does not exist.'));
+
 				if(bccomp(self::$userData['userid'], $user['userid']) != 0){
 					if(USER_TYPE_SUPER_ADMIN != self::$userData['type'])
 						self::exception(ZBX_API_ERROR_PARAMETERS, _s('You do not have permissions to update other users.'));
@@ -380,6 +392,12 @@ Copt::memoryPick();
 				$dbUser = $dbUsers[$user['userid']];
 			}
 			else{
+				if(USER_TYPE_SUPER_ADMIN != self::$userData['type'])
+					self::exception(ZBX_API_ERROR_PARAMETERS, _s('You do not have permissions to delete users.'));
+
+				if(!isset($dbUsers[$user['userid']]))
+					self::exception(ZBX_API_ERROR_PARAMETERS, _s('You do not have permissions to delete user or user does not exist.'));
+
 				if(bccomp(self::$userData['userid'], $user['userid']) == 0)
 					self::exception(ZBX_API_ERROR_PARAMETERS, _s('User is not allowed to delete himself.'));
 
@@ -568,8 +586,8 @@ Copt::memoryPick();
 
 			$result = DB::update('users', array(
 				array(
-						'values'=> $user,
-						'where'=> array('userid='.$user['userid'])
+					'values'=> $user,
+					'where'=> array('userid' => $user['userid'])
 				)
 			));
 
@@ -989,37 +1007,39 @@ Copt::memoryPick();
 	public function checkAuthentication($sessionid){
 		global $ZBX_LOCALNODEID;
 
-		$sql = 'SELECT u.userid, u.autologout'.
+		$time = time();
+		$sql = 'SELECT u.userid, u.autologout, s.lastaccess '.
 				' FROM sessions s, users u'.
 				' WHERE s.sessionid='.zbx_dbstr($sessionid).
 					' AND s.status='.ZBX_SESSION_ACTIVE.
 					' AND s.userid=u.userid'.
-					' AND ((s.lastaccess+u.autologout>'.time().') OR (u.autologout=0))'.
+					' AND ((s.lastaccess+u.autologout>'.$time.') OR (u.autologout=0))'.
 					' AND '.DBin_node('u.userid', $ZBX_LOCALNODEID);
 		$userInfo = DBfetch(DBselect($sql));
 
 		if(!$userInfo)
 			self::exception(ZBX_API_ERROR_PARAMETERS, _('Session terminated, re-login, please.'));
 
-		if(!check_perm2system($userInfo['userid']))
-			self::exception(ZBX_API_ERROR_PARAMETERS, _('No permissions for system access.'));
+// dont check permissions on the same second
+		if($time != $userInfo['lastaccess']){
+			if(!check_perm2system($userInfo['userid']))
+				self::exception(ZBX_API_ERROR_PARAMETERS, _('No permissions for system access.'));
 
-		if($userInfo['autologout'] > 0)
-			DBexecute('DELETE FROM sessions WHERE userid='.$userInfo['userid'].' AND lastaccess<'.(time() - $userInfo['autologout']));
+			if($userInfo['autologout'] > 0)
+				DBexecute('DELETE FROM sessions WHERE userid='.$userInfo['userid'].' AND lastaccess<'.(time() - $userInfo['autologout']));
 
-		DBexecute('UPDATE sessions SET lastaccess='.time().' WHERE userid='.$userInfo['userid'].' AND sessionid='.zbx_dbstr($sessionid));
+			DBexecute('UPDATE sessions SET lastaccess='.time().' WHERE userid='.$userInfo['userid'].' AND sessionid='.zbx_dbstr($sessionid));
+		}
 
 		$sql = 'SELECT MAX(g.gui_access) as gui_access '.
 			' FROM usrgrp g, users_groups ug '.
 			' WHERE ug.userid='.$userInfo['userid'].
 				' AND g.usrgrpid=ug.usrgrpid ';
 		$db_access = DBfetch(DBselect($sql));
-		if(!zbx_empty($db_access['gui_access'])){
+		if(!zbx_empty($db_access['gui_access']))
 			$guiAccess = $db_access['gui_access'];
-		}
-		else{
+		else
 			$guiAccess = GROUP_GUI_ACCESS_SYSTEM;
-		}
 
 		$userData = $this->_getUserData($userInfo['userid']);
 		$userData['sessionid'] = $sessionid;
